Move buffer from digitizing to selection
