+++ lizmap/modules/view/locales/en_US/dictionnary.UTF-8.properties
@@ -145,6 +145,7 @@ toolbar.content.stop=Stop
 
 selectiontool.toolbar.title=Selection
 selectiontool.toolbar.layer=Selection layer
+selectiontool.toolbar.buffer=Buffer
 selectiontool.toolbar.action.type.refresh=New selection
 selectiontool.toolbar.action.type.plus=Add to selection
 selectiontool.toolbar.action.type.minus=Remove from selection
@@ -161,7 +162,6 @@ digitizing.toolbar.query.point=Point
 digitizing.toolbar.query.line=Line
 digitizing.toolbar.query.polygon=Polygon
 digitizing.toolbar.query.freehand=Free hand
-digitizing.toolbar.buffer=Buffer
 
 switcher.layer.export.title=Export
